Fix forward pass precision and reduce memory for training
- cross_entropy_fwd: use log-sum-exp directly instead of log(softmax+eps),
  eliminating the 1e-10 epsilon and improving numerical precision. Loss diff
  vs PyTorch now 2.3e-05 (was 1.05e-03).
- model.c/h: change decay_buf from float* to uint8_t*, saving 372 MB for
  GPT-2 scale (4 vs 5 float32 buffers). Enables training with B=1..2 on a
  machine with ~3 GB RAM.
- verify_forward.py: use PyTorch float32 as primary reference when available,
  fall back to numpy. Update default tolerance to 5e-4 (expected float32
  accumulation for a 12-layer model with different BLAS backends).
- autotools: add generated configure, Makefile.in, and helper files so users
  can build without installing autoconf/automake.

+++ src/model.h
@@ -11,6 +11,7 @@
 #define CNGPT_MODEL_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 /* -----------------------------------------------------------------------
  * Weight file header (7 × int32)
@@ -130,7 +131,7 @@ typedef struct {
     /* AdamW optimizer state */
     float *m_buf;                 /* first moment,  same size as param_buf */
     float *v_buf;                 /* second moment, same size as param_buf */
-    float *decay_buf;             /* 1.0 = apply weight decay, 0.0 = skip (biases/LN) */
+    uint8_t *decay_buf;           /* 1 = apply weight decay, 0 = skip (biases/LN) */
     int    n_params;              /* total parameter count */
     int    step;                  /* optimizer step counter */
 } GPT;
